move socket to project related component to avoid undefined user
